Manage dependencies with plugin bill of materials
The plugin bill of materials allows the plugin maintainer to rely on
the plugin bom to provide version numbers for many common
dependencies.  That avoids effort trying to find a set of version
numbers that combine well for testing and simplifies plugin
maintenance.

See https://github.com/jenkinsci/bom#readme for more details on the
ways that plugin bill of materials can help.