Use okhttp-api Jenkins plugin instead of library
Switch the okhttp dependency from bundling the okhttp jar file inside
the plugin to use the Jenkins okhttp-api plugin that is already
installed on many Jenkins controllers.  Reduces the size of the plugin
and assures that the okhttp library version is consistent with the
library used in other Jenkins plugins.